Solved the occlusion problem and scaling problem in the map page.

+++ src/com/andyidea/tabdemo/image/DragImageView.java
@@ -144,8 +144,8 @@ public class DragImageView extends ImageView {
 //		MAX_W = bitmap_W * 3;
 //		MAX_H = bitmap_H * 3;
 
-		MAX_W = (int) (bitmap_W * 5);
-		MAX_H = (int) (bitmap_H * 5);
+		MAX_W = bitmap_W * 5;
+		MAX_H = bitmap_H * 5;
 
 		MIN_W = bitmap_W;
 		MIN_H = bitmap_H;
